requirements.txt
streamlit
pandas
python-IO
datetime
python-dotenv
google.cloud.storage
python-time
google-cloud-aiplatform==1.36.3
langchain==0.0.323
trulens-eval 
litellm==0.14.1
pydub
llama-index
trulens
python-utils
transformers
torch
sentence-transformers


